Fixed bug involving extreme cell changes
Fixed a bug where assumptions in the cell list could
be invalidated on a very large cell change, leading
to a panic. Did this by allowing apply_random_strain
to fail gracefully, assuming that if it does, the configuration
would be invalid anyway. Prints warning, since this should be
extremely rare in practice, and indicates that bad parameters
for the run were selected. Also fixed a bug where exact steps in
ln V were larger than envisioned, since I forgot to include a power
of 1/d in the expression for the cell transformation.