Added new ObservableTracker struct
This allows data of arbitrary type to be
used in computing observables over the run.
Currently do not specialize by P or C, because
the struct is only fully initialized after each sweep
at the moment, meaning it is not performance critical, and
we can keep around every field needed for every particle type.